[FLINK-13502][table-planner-blink] Move CatalogTableStatisticsConverter & TreeNode to correct package
This closes #9284
