+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/expressions/PlannerExpression.scala
@@ -19,7 +19,6 @@ package org.apache.flink.table.planner.expressions
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.table.expressions.{Expression, ExpressionVisitor}
-import org.apache.flink.table.planner.plan.TreeNode
 import org.apache.flink.table.planner.validate.{ValidationResult, ValidationSuccess}
 
 import java.util
